React functional components, react scripts in package.json and assignment for creating header component
