+++ index.html
@@ -9,7 +9,7 @@
 
 <body>
     <div id="root"></div>
-    <script type="module" src="./index.js"></script>
+    <script type="module" src="./App.js"></script>
 </body>
 
 </html>
