+++ package.json
@@ -2,8 +2,9 @@
   "name": "namastereact",
   "version": "1.0.0",
   "description": "React with parcel",
-  
   "scripts": {
+    "start": "parcel index.html",
+    "build": "parcel build index.html",
     "test": "jest"
   },
   "repository": {
@@ -21,9 +22,10 @@
   },
   "dependencies": {
     "react": "^19.1.0",
-    "react-dom": "^19.1.0"
+    "react-dom": "^19.1.0",
+    "react-icons": "^5.5.0"
   },
-  "browsersList" : [
+  "browsersList": [
     "last 2 version"
   ]
 }
